Move suggested-fix logic into CrashSuggestedFix class
Extract GetSuggestedFix/plugin-crash detection out of Main.cs into a dedicated
CrashSuggestedFix class so the list of known crashes can grow without cluttering
Bootstrap. Cast the exception once instead of repeatedly, and reword the plugin
fix note (drop the assistant/copilot aside, use 'UO').
